Extract shared contact methods into base Contact class
